Your commit message

+++ delete.php
@@ -9,10 +9,18 @@ if (isset($_GET["id"])) {
 
     $conn = new mysqli($servername, $username, $password, $database);
 
+    if ($conn->connect_error) {
+        die("Connection failed: " . $conn->connect_error);
+    }
+
     $sql = "DELETE FROM homeowners WHERE id = $id";
-    $conn->query($sql);
+    if ($conn->query($sql) === TRUE) {
+        echo "Record deleted successfully";
+    } else {
+        echo "Error deleting record: " . $conn->error;
+    }
 
-
+    $conn->close();
 }
 header("location: homeowneradmin.php");
 exit;
